requirements.txt
openai==1.46.0
requests
python-dotenv
pydantic==2.9.1
instructor==1.4.3